+++ Exercises/CheckIfItIsAStraightLine.cpp
@@ -1,2 +1,8 @@
-#include "CheckIfItIsAStraightLine.h"
-//TODO
+#include <vector>
+
+class Solution
+{
+public:
+	bool checkStraightLine(std::vector<std::vector<int>>& coordinates);
+	//TODO
+};
