--- Exercises/CheckIfItIsAStraightLine.h
@@ -1,11 +1,0 @@
-#pragma once
-#include <vector>
-
-namespace CheckIfItIsAStraightLine
-{
-	class Solution
-	{
-	public:
-		bool checkStraightLine(std::vector<std::vector<int>>& coordinates);
-	};
-}
