+++ Exercises/CousinsInBinaryTree.cpp
@@ -1,2 +1,8 @@
-#include "CousinsInBinaryTree.h"
-//TODO
+#include "TreeNode.h"
+
+class Solution
+{
+public:
+	bool isCousins(TreeNode* root, int x, int y);
+	//TODO
+};
