--- Exercises/CousinsInBinaryTree.h
@@ -1,11 +1,0 @@
-#pragma once
-#include "TreeNode.h"
-
-namespace CounsinsInBinaryTree
-{
-	class Solution
-	{
-	public:
-		bool isCousins(TreeNode* root, int x, int y);
-	};
-}
