--- Exercises/FindAllAnagramsInAString.h
@@ -1,12 +1,0 @@
-#pragma once
-#include <string>
-#include <vector>
-
-namespace FindAllAnagramsInAString
-{
-	class Solution
-	{
-	public:
-		std::vector<int> findAnagrams(std::string s, std::string p);
-	};
-}
