+++ Exercises/FindTheTownJudge.cpp
@@ -1,2 +1,8 @@
-#include "FindTheTownJudge.h"
-//TODO
+#include <vector>
+
+class Solution
+{
+public:
+	int findJudge(int N, std::vector<std::vector<int>>& trust);
+	//TODO
+};
