--- Exercises/FindTheTownJudge.h
@@ -1,11 +1,0 @@
-#pragma once
-#include <vector>
-
-namespace FindTheTownJudge
-{
-	class Solution
-	{
-	public:
-		int findJudge(int N, std::vector<std::vector<int>>& trust);
-	};
-}
