+++ Exercises/FirstBadVersion.cpp
@@ -1,10 +1,8 @@
-#include "FirstBadVersion.h"
-
-namespace FirstBadVersion
+class Solution
 {
-	Solution::Solution(int version) : mpFirstBadVersion(version) {}
-
-	int Solution::firstBadVersion(int n)
+public:
+	Solution(int version) : mpFirstBadVersion(version) {}
+	int firstBadVersion(int n)
 	{
 		unsigned int start = 1;
 		unsigned int end = n;
@@ -21,8 +19,10 @@ namespace FirstBadVersion
 		return end;
 	}
 
-	bool Solution::isBadVersion(int version)
+private:
+	int mpFirstBadVersion;
+	bool isBadVersion(int version)
 	{
 		return (version >= mpFirstBadVersion);
 	}
-}
+};
