+++ Exercises/FirstUniqueCharacterInAString.cpp
@@ -1,2 +1,8 @@
-#include "FirstUniqueCharacterInAString.h"
-//TODO
+#include <string>
+
+class Solution
+{
+public:
+	int firstUniqChar(std::string s);
+	//TODO
+};
