--- Exercises/FirstUniqueCharacterInAString.h
@@ -1,11 +1,0 @@
-#pragma once
-#include <string>
-
-namespace FirstUniqueCharacterInAString
-{
-	class Solution
-	{
-	public:
-		int firstUniqChar(std::string s);
-	};
-}
