--- Exercises/JewelsAndStones.h
@@ -1,11 +1,0 @@
-#pragma 
-#include <string>
-
-namespace JewelsAndStones
-{
-	class Solution
-	{
-	public:
-		int numJewelsInStones(std::string J, std::string S);
-	};
-}
