--- Exercises/MajorityElement.h
@@ -1,11 +1,0 @@
-#pragma once
-#include <vector>
-
-namespace MajorityElement
-{
-	class Solution
-	{
-	public:
-		int majorityElement(std::vector<int>& nums);
-	};
-}
