+++ Exercises/MaximumSumCircularSubarray.cpp
@@ -1,2 +1,8 @@
-#include "MaximumSumCircularSubarray.h"
-//TODO
+#include <vector>
+
+class Solution
+{
+public:
+	int maxSubarraySumCircular(std::vector<int>& A);
+	//TODO
+};
