--- Exercises/MaximumSumCircularSubarray.h
@@ -1,11 +1,0 @@
-#pragma once
-#include <vector>
-
-namespace MaximumSumCircularSubarray
-{
-	class Solution
-	{
-	public:
-		int maxSubarraySumCircular(std::vector<int>& A);
-	};
-}
