--- Exercises/NumberComplement.h
@@ -1,10 +1,0 @@
-#pragma once
-
-namespace NumberComplement
-{
-	class Solution
-	{
-	public:
-		int findComplement(int num);
-	};
-}
