--- Exercises/OddEvenLinkedList.h
@@ -1,11 +1,0 @@
-#pragma once
-#include "ListNode.h"
-
-namespace OddEvenLinkedList
-{
-	class Solution
-	{
-	public:
-		ListNode* oddEvenList(ListNode* head);
-	};
-}
