+++ Exercises/RansomNote.cpp
@@ -1,2 +1,8 @@
-#include "RansomNote.h"
-//TODO
+#include <string>
+
+class Solution
+{
+public:
+	bool canConstruct(std::string ransomNote, std::string magazine);
+	//TODO
+};
