--- Exercises/RansomNote.h
@@ -1,11 +1,0 @@
-#pragma once
-#include <string>
-
-namespace RansomNote
-{
-	class Solution
-	{
-	public:
-		bool canConstruct(std::string ransomNote, std::string magazine);
-	};
-}
