--- Exercises/RemoveKDigits.h
@@ -1,11 +1,0 @@
-#pragma once
-#include <string>
-
-namespace RemoveKDigits
-{
-	class Solution
-	{
-	public:
-		std::string removeKdigits(std::string num, int k);
-	};
-}
