--- Exercises/SingleElementInASortedArray.h
@@ -1,11 +1,0 @@
-#pragma once
-#include <vector>
-
-namespace SingleElementInASortedArray
-{
-	class Solution
-	{
-	public:
-		int singleNonDuplicate(std::vector<int>& nums);
-	};
-}
