--- Exercises/ValidPerfectSquare.h
@@ -1,10 +1,0 @@
-#pragma once
-
-namespace ValidPerfectSquare
-{
-	class Solution
-	{
-	public:
-		bool isPerfectSquare(int num);
-	};
-}
