+++ Tests/FindAllAnagramsInAStringTest.cpp
@@ -1,1 +1,8 @@
 #include "catch.hpp"
+#include "FindAllAnagramsInAString.cpp"
+
+TEST_CASE("Test")
+{
+	Solution s;
+	s.findAnagrams("a", "a");
+}
