+++ Tests/JewelsAndStonesTest.cpp
@@ -1,19 +1,16 @@
 #include "catch.hpp"
-#include "JewelsAndStones.h"
+#include "JewelsAndStones.cpp"
 
 #include <string>
 
-namespace JewelsAndStones
+TEST_CASE("Example 1")
 {
-	TEST_CASE("Example 1")
-	{
-		Solution s;
-		REQUIRE(3 == s.numJewelsInStones("aA", "aAAbbbb"));
-	}
+	Solution s;
+	REQUIRE(3 == s.numJewelsInStones("aA", "aAAbbbb"));
+}
 
-	TEST_CASE("Example 2")
-	{
-		Solution s;
-		REQUIRE(0 == s.numJewelsInStones("z", "ZZ"));
-	}
+TEST_CASE("Example 2")
+{
+	Solution s;
+	REQUIRE(0 == s.numJewelsInStones("z", "ZZ"));
 }
