+++ Tests/Tests.vcxproj.filters
@@ -18,5 +18,11 @@
     <ClCompile Include="FindAllAnagramsInAStringTest.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="ImplementTrieTest.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="JewelsAndStonesTest.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
 </Project>
